semantic_knowledge_base_project/requirements.txt
sentence-transformers
faiss-cpu
numpy
PyPDF2
python-docx
pymupdf
